+++ plugins/api/src/getRedirects.js
@@ -1,13 +1,22 @@
-const { normalize, defaults } = require('./_fns')
+const { normalize, defaults, toUrlParams } = require('./_fns')
+
+const params = {
+  action: 'query',
+  rdnamespace: 0,
+  prop: 'redirects',
+  rdlimit: 500,
+  format: 'json',
+  origin: '*',
+  redirects: true
+}
 
 const makeUrl = function (title, options, cursor) {
-  title = normalize(title)
-  title = encodeURIComponent(title)
   let url = `https://${options.lang}.wikipedia.org/${options.path}?`
   if (options.domain) {
     url = `https://${options.domain}/${options.path}?`
   }
-  url += `action=query&titles=${title}&rdnamespace=0&prop=redirects&rdlimit=500&format=json&origin=*&redirects=true`
+  url += toUrlParams(params)
+  url += `&titles=${normalize(title)}`
   if (cursor) {
     url += '&rdcontinue=' + cursor
   }